Add opt-in unit-scale FP8 MiMo cache and prefix storage

+++ src/models/mimo/model.hpp
@@ -100,8 +100,8 @@ class MimoModel : public SessionModel<MimoModel> {
   MimoGlobalsResident globals_;
   std::vector<MimoLayerResident> weights_;
   LayerBump cache_, scratch_, attention_scores_, layer_workspace_;
-  std::vector<uint16_t*> keys_, values_;
-  std::vector<size_t> key_plane_, value_plane_;
+  std::vector<uint8_t*> keys_, values_;
+  std::vector<size_t> key_plane_, value_plane_;  // bytes per request, unlike BF16 draft planes
   std::unique_ptr<GlmMoeLayer> shared_moe_;
   std::vector<std::unique_ptr<MimoDecoderLayer>> layers_;
   uint16_t* residual_ = nullptr;
